refactor: DimiRu ButtonGroup-related components (#186)
* refactor: DimiRu ButtonGroup-related components

* fix: add name about DimiButtonGroup

* fix: wrong imports in DimiRu

* style: change to directly declaring functions
