+++ app/dimiru/components/DimiButtonGroup.vue
@@ -1,5 +1,7 @@
 <script>
 export default {
+  name: 'DimiButtonGroup',
+
   props: {
     value: {
       type: Number,
@@ -9,24 +11,6 @@ export default {
     items: {
       type: Array,
       required: true
-    },
-
-    colors: {
-      type: Array,
-      default: () => ['gray']
-    }
-  },
-
-  computed: {
-    color () {
-      return this.colors[this.value]
-    },
-
-    classes () {
-      return [
-        'button-group',
-        `button-group--${this.color}`
-      ]
     }
   },
 
@@ -48,12 +32,12 @@ export default {
 </script>
 
 <template>
-  <div :class="classes">
+  <div class="long-button-group">
     <div
       v-for="(item, index) in items"
       :key="`button-${index}`"
       :active="index === value"
-      class="button-group__button"
+      class="long-button-group__button"
       @click="onClick(index)"
     >
       {{ item }}
@@ -64,18 +48,18 @@ export default {
 <style lang="scss">
 @import '../scss/vars';
 
-.button-group {
+.long-button-group {
   display: inline-block;
-  font-size: 14px;
+  color: $white;
+  font-size: 18px;
   font-weight: $font-weight-bold;
   user-select: none;
 
   &__button {
     display: inline-block;
-    width: 70px;
-    padding-top: 0.35em;
-    padding-bottom: 0.25em;
-    border: 1px solid $black;
+    width: 205px;
+    padding: 0.7em 2em;
+    background-color: $red;
     cursor: pointer;
     text-align: center;
     transition: all 0.2s ease-in-out;
@@ -92,38 +76,7 @@ export default {
   }
 
   &__button:not(:last-child) {
-    border-right: 0;
+    border-right: 1px solid rgba(255, 255, 255, 0.3);
   }
 }
-
-@mixin colored($color, $value) {
-  .button-group--#{$color} {
-    .button-group__button {
-      border-color: $value;
-      color: $value;
-    }
-
-    .button-group__button[active] {
-      background-color: $value;
-      color: $white;
-    }
-  }
-}
-
-@include colored('white', $white);
-@include colored('black', $black);
-@include colored('gray', $gray);
-@include colored('gray-dark', $gray-dark);
-@include colored('gray-light', $gray-light);
-@include colored('gray-lighter', $gray-lighter);
-@include colored('gray-lighten', $gray-lighten);
-
-@include colored('red', $red);
-@include colored('red-dark', $red-dark);
-
-@include colored('pink', $pink);
-@include colored('cyan', $cyan);
-@include colored('aloes', $aloes);
-@include colored('orange', $orange);
-@include colored('mustard', $mustard);
 </style>
