+++ app/dimiru/components/index.js
@@ -1,7 +1,8 @@
 export { default as DimiBadge } from './DimiBadge.vue'
+export { default as DimiBadgeGroup } from './DimiBadgeGroup.vue'
 export { default as DimiButton } from './DimiButton.vue'
-export { default as DimiButtonRadio } from './DimiButtonRadio.vue'
 export { default as DimiButtonGroup } from './DimiButtonGroup.vue'
+export { default as DimiButtonRadio } from './DimiButtonRadio.vue'
 export { default as DimiCard } from './DimiCard.vue'
 export { default as DimiCheckbox } from './DimiCheckbox.vue'
 export { default as DimiDateInput } from './DimiDateInput.vue'
@@ -17,4 +18,3 @@ export { default as DimiModal } from './DimiModal.vue'
 export { default as DimiNavbar } from './DimiNavbar.vue'
 export { default as DimiTab } from './DimiTab.vue'
 export { default as DimiLongInput } from './DimiLongInput.vue'
-export { default as DimiLongButtonGroup } from './DimiLongButtonGroup.vue'
